fix(serial_frame): make sure frame size is greater than 0 when decoding
Otherwise, an incoming payload with only a START_BYTE and an END_BYTE
will cause the decoder to perform an out of bound read on the frame
buffer causing a segmentation fault.
This was due to the fact that we relied on the frame offset to compute
the framesize. In this case the frame offset when processing the
END_BYTE is 0. When we compute the framesize, we substract the expected
CRC from the frame: framesize = decoder->frame_offset - 1. In this
particular case, we causes an underflow on framesize which is unsigned.
And then passing it to following functions will causes the out-of-bounds
read.

This commit also add a test for this particular case.