Introduce SmpContext to replace SmpSerialFrameContext
and hide implementation details of the context. The goal is to
completely replace the SmpSerialFrameContext object which is quite low
level and need buffer size configured at compile time.
From user point of view, the main difference is the usage of new and free
function to create and destroy a context and the message callback now
directly give a SmpMessage instead of the serial frame.
For now, the serial decoder buffer is kept to its default value, 1kB,
without any way to change it.