serial-protocol: reset decoder state on error
to not continue to decode a frame we can't finish and throw continuous
error. This bug was revealed by the context static api check. Indeed
when playing with a too little rx buffer, each byte which exceed the
decoder capacity causes a too big error.
So in error cases, just wait for the next start byte, unless the error
is a start byte within a frame.