message: dynamically allocate values array instead of a fixed configuration
This commit remove the compile time option limiting the number of values within
a message and use malloc to allocate them when the message is created.
The default capacity of a SmpMessage is 8 by default and can be increased by
calling smp_message_set_capacity(). As a consequence and to avoid any mistakes,
smp_message_init() and smp_message_init_from_buffer() functions have been
removed.